requirements.txt
flask
flask-caching
flask-sqlalchemy
pymysql
python-dotenv
slack-sdk
openai
anthropic
google-cloud-language
slack_bolt
slack_sdk